src/ai_manager.py
import os
import json
import google.generativeai as genai
from PIL import Image
from io import BytesIO
import requests
# rembg needs to be handled carefully as it might not be installed in some environments
try:
    from rembg import remove
    REMBG_AVAILABLE = True
except ImportError:
    REMBG_AVAILABLE = False
    print("Warning: rembg not found. Background removal disabled.")

from src.config import GEMINI_API_KEY, GEMINI_MODEL_TEXT, ASSETS_PATH

# Configure Gemini
if GEMINI_API_KEY:
    genai.configure(api_key=GEMINI_API_KEY)

class AIManager:
    def __init__(self):
        self.model_text = genai.GenerativeModel(GEMINI_MODEL_TEXT)
        # Note: Image generation usually requires a specific client or endpoint in Vertex AI
        # or the specific Gemini multimodal capability.
        # For this 'free tier' request, we assume the standard GenerativeModel usage if available
        # or fallback to text-based description if image gen isn't directly accessible via this SDK
        # in the same way.
        # However, Gemini 2.5 / Imagen 3 access is requested.
        # Since I cannot guarantee the specific 'imagen' endpoint here,
        # I will structure the code to use a hypothetical image generation method
        # or a standard one if the SDK supports it.
        pass

    def generate_monster_stats(self, level=1, context="random"):
        """
        Generates JSON stats for a new monster.
        """
        prompt = f"""
        Create a unique RPG monster inspired by Pokémon.
        Context: {context}. Level: {level}.
        Return ONLY a valid JSON object with the following structure, no markdown formatting:
        {{
            "name": "Name",
            "is_mythical": false,
            "type_1": "Fire",
            "type_2": "null or Type",
            "hp_max": 50,
            "mp_max": 20,
            "attack": 10,
            "defense": 10,
            "speed": 10,
            "description": "Short visual description for image generation"
        }}
        For mythical monsters (1% chance usually handled by caller, but you can suggest one), stats should be higher.
        Base stats should be appropriate for Level {level}.
        """

        try:
            response = self.model_text.generate_content(prompt)
            # Clean response of markdown code blocks if present
            text = response.text.strip()
            if text.startswith("```json"):
                text = text[7:]
            if text.endswith("```"):
                text = text[:-3]
            return json.loads(text)
        except Exception as e:
            print(f"Error generating monster stats: {e}")
            # Fallback
            return {
                "name": "Glitch",
                "is_mythical": False,
                "type_1": "Normal",
                "type_2": None,
                "hp_max": 20,
                "mp_max": 10,
                "attack": 5,
                "defense": 5,
                "speed": 5,
                "description": "A glitchy pixelated blob."
            }

    def generate_abilities(self, monster_type, count=5):
        """
        Generates a list of abilities.
        """
        prompt = f"""
        Create {count} RPG abilities for a monster of type {monster_type}.
        Return ONLY a valid JSON array of objects, no markdown:
        [
            {{
                "name": "Ability Name",
                "description": "Effect description",
                "type": "{monster_type}",
                "damage": 10,
                "heal": 0,
                "cost_mp": 5,
                "cost_hp": 0,
                "cooldown_local": 0,
                "cooldown_global": 0,
                "stun_duration": 0,
                "drain_percent": 0,
                "is_legendary": false,
                "visual_description": "Description for image generation"
            }}
        ]
        """
        try:
            response = self.model_text.generate_content(prompt)
            text = response.text.strip()
            if text.startswith("```json"):
                text = text[7:]
            if text.endswith("```"):
                text = text[:-3]
            return json.loads(text)
        except Exception as e:
            print(f"Error generating abilities: {e}")
            return []

    def generate_image(self, description, filename_prefix, is_monster=True):
        """
        Generates an image (mocked or using available tools), removes background, and saves it.
        Since I cannot verify the Image Generation API quota or access in this environment,
        I will create a placeholder image if generation fails or isn't configured.
        """

        # TODO: Insert actual Gemini Image Generation call here if available.
        # Currently, the standard google-generativeai python lib focuses on text/multimodal input.
        # Image generation (Imagen) often uses a different client or REST API.
        # For this prototype, I will create a placeholder or attempt a text-to-image if the model supports it.

        # Placeholder Logic for Sandbox:
        # Create a colored square based on description hash or random.
        image_path = os.path.join(ASSETS_PATH, f"{filename_prefix}.png")

        # In a real scenario, we would do:
        # 1. Call API -> Get Image URL or Bytes
        # 2. img = Image.open(BytesIO(response.content))
        # 3. if REMBG_AVAILABLE: img = remove(img)
        # 4. img.save(image_path)

        self._create_placeholder_image(image_path, description)
        return image_path

    def _create_placeholder_image(self, path, description):
        # Create a simple colored block with text
        from PIL import Image, ImageDraw
        import random

        img = Image.new('RGBA', (256, 256), (0, 0, 0, 0))
        draw = ImageDraw.Draw(img)

        # Random color blob
        color = (random.randint(50, 255), random.randint(50, 255), random.randint(50, 255), 255)
        draw.ellipse([20, 20, 236, 236], fill=color)

        img.save(path)

    def evolve_monster_stats(self, current_stats, evolution_stage):
        """
        Evolves the stats and name.
        """
        multiplier_min = 5 if evolution_stage == 1 else 10
        multiplier_max = 15 if evolution_stage == 1 else 25

        prompt = f"""
        Evolve this monster: {json.dumps(current_stats)}.
        Evolution Stage: {evolution_stage} (going to {evolution_stage+1}).
        Increase stats by random % between {multiplier_min}-{multiplier_max}%.
        Create a new cooler name based on the old one.
        Return ONLY JSON:
        {{
            "name": "New Name",
            "hp_max": 100,
            "mp_max": 40,
            "attack": 20,
            "defense": 20,
            "speed": 20,
            "description": "Description of evolved form"
        }}
        """
        try:
            response = self.model_text.generate_content(prompt)
            text = response.text.strip()
            if text.startswith("```json"):
                text = text[7:]
            if text.endswith("```"):
                text = text[:-3]
            return json.loads(text)
        except Exception as e:
            print(f"Error generating evolution: {e}")
            return current_stats # Fail safe
