requirements.txt
PyQt6
google-generativeai
pillow
rembg
cryptography
requests
